+++ NEWS.md
@@ -4,6 +4,7 @@
 * Developed AT from: 2024-02-01,
 * add crf_status in patientsurvival
 * add transplantationsurvival
+* add categorize_pathogenspecies()
 
 
 # stcsanalysis 0.1.0